% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_rows.R
\name{add_body_row}
\alias{add_body_row}
\title{Add body labels}
\usage{
add_body_row(x, top = TRUE, values = list(), colwidths = integer(0))
}
\arguments{
\item{x}{a flextable object}

\item{top}{should the row be inserted at the top or the bottom.}

\item{values}{values to add (as a list).}

\item{colwidths}{the number of columns to merge in the row for each label}
}
\description{
Add a row of new columns labels in body part.
Labels can be spanned along multiple columns, as merged cells.

Labels are associated with a number of columns
to merge that default to one if not specified.
In this case, you have to make sure that the
number of labels is equal to the number of columns displayed.

The function can add only one single row by call.
}
\examples{
ft <- flextable(head(iris))
ft <- add_body_row(ft, values = list(1000), colwidths = 5)
ft
}
\seealso{
\code{\link[=flextable]{flextable()}}, \code{\link[=add_header_row]{add_header_row()}}

Other functions that add lines in the table: 
\code{\link{add_body}()},
\code{\link{add_footer_lines}()},
\code{\link{add_footer_row}()},
\code{\link{add_footer}()},
\code{\link{add_header_row}()},
\code{\link{add_header}()}
}
\concept{functions that add lines in the table}
