% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{add_header}
\alias{add_header}
\title{Add a row of labels in headers}
\usage{
add_header(x, top = TRUE, ...)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{top}{should the row be inserted at the top or the bottom.}

\item{...}{a named list (names are data colnames) of strings
specifying corresponding labels to add.}
}
\description{
Add a single row of labels in the flextable's header part. It can
be inserted at the top or the bottom of header part.
}
\details{
An horizontal and a vertical merge are performed automatically.
Use \code{\link{merge_none}} to drop theses merging instructions.
}
\examples{
ft <- flextable( head( iris ),
  col_keys = c("Species", "Sepal.Length", "Petal.Length", "Sepal.Width", "Petal.Width") )
ft <- add_header(x = ft, Sepal.Length = "length",
  Sepal.Width = "width", Petal.Length = "length",
  Petal.Width = "width", Species = "Species", top = FALSE )
ft <- add_header(ft, Sepal.Length = "Inches",
  Sepal.Width = "Inches", Petal.Length = "Inches",
  Petal.Width = "Inches", Species = "Species", top = TRUE )
ft <- merge_h(ft, part = "header")
ft <- autofit(ft)
write_docx("ft_add_header.docx", ft)
}

