% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_str.R
\name{add_latex_dep}
\alias{add_latex_dep}
\title{add latex dependencies}
\usage{
add_latex_dep(float = FALSE, wrapfig = FALSE)
}
\arguments{
\item{float}{load package 'float'}

\item{wrapfig}{load package 'wrapfig'}
}
\description{
Manually add flextable latex dependencies to
the knitr session via \code{\link[=knit_meta_add]{knit_meta_add()}}.

When enabling caching in 'R Markdown' documents for PDF output,
the flextable cached result is used directly. Call \code{add_latex_dep()} in a
non cached chunk so that flextable latex dependencies are added
to knitr metadata.
}
\examples{
add_latex_dep()
}
