% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-selectors.R
\name{args_x_j}
\alias{args_x_j}
\title{internal utils for roxygen tags reuse}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}

\item{j}{column selector, see section \emph{Column selection with the \code{j} parameter}
in <\code{\link[=flextable_selectors]{Selectors in flextable}}>.}
}
\description{
internal utils for roxygen tags reuse
}
\keyword{internal}
