% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc-selectors.R
\name{args_x_only}
\alias{args_x_only}
\title{internal utils for roxygen tags reuse}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}
}
\description{
internal utils for roxygen tags reuse
}
\keyword{internal}
