% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.glm}
\alias{as_flextable.glm}
\title{tabular summary for lm object}
\usage{
\method{as_flextable}{glm}(x, ...)
}
\arguments{
\item{x}{glm model}

\item{...}{unused argument}
}
\description{
produce a flextable describing a
generalized linear model produced by function \code{glm}.
}
\examples{
if(require("broom")){
  dat <- attitude
  dat$high.rating <- (dat$rating > 70)
  probit.model <- glm(high.rating ~ learning + critical +
     advance, data=dat, family = binomial(link = "probit"))
  ft <- as_flextable(probit.model)
  ft
}
}
