% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.gt_tbl.R
\name{as_flextable.gt_tbl}
\alias{as_flextable.gt_tbl}
\title{Transform a 'gt' object into a flextable}
\usage{
\method{as_flextable}{gt_tbl}(x, ...)
}
\arguments{
\item{x}{a \code{\link[gt:gt]{gt::gt()}} object}

\item{...}{unused argument}
}
\description{
produce a flextable from a 'gt_tbl' object.

This function extracts:
\enumerate{
\item The underlying data frame.
\item The column labels (\code{cols_label}).
\item The spanner column labels (\code{tab_spanner}).
}
}
\seealso{
Other as_flextable methods: 
\code{\link{as_flextable}()},
\code{\link{as_flextable.data.frame}()},
\code{\link{as_flextable.gam}()},
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.kmeans}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.merMod}()},
\code{\link{as_flextable.pam}()},
\code{\link{as_flextable.summarizor}()},
\code{\link{as_flextable.table}()},
\code{\link{as_flextable.tabular}()},
\code{\link{as_flextable.tabulator}()},
\code{\link{as_flextable.xtable}()}
}
\concept{as_flextable methods}
