% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.lm}
\alias{as_flextable.lm}
\title{tabular summary for lm object}
\usage{
\method{as_flextable}{lm}(x, ...)
}
\arguments{
\item{x}{lm model}

\item{...}{unused argument}
}
\description{
produce a flextable describing a
linear model produced by function \code{lm}.
}
\examples{
if(require("broom")){
  lmod <- lm(rating ~ complaints + privileges +
    learning + raises + critical, data=attitude)
  as_flextable(lmod)
}
}
