% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.merMod}
\alias{as_flextable.merMod}
\alias{as_flextable.lme}
\alias{as_flextable.gls}
\alias{as_flextable.nlme}
\alias{as_flextable.brmsfit}
\alias{as_flextable.glmmTMB}
\alias{as_flextable.glmmadmb}
\title{tabular summary for mixed model}
\usage{
\method{as_flextable}{merMod}(x, ...)

\method{as_flextable}{lme}(x, ...)

\method{as_flextable}{gls}(x, ...)

\method{as_flextable}{nlme}(x, ...)

\method{as_flextable}{brmsfit}(x, ...)

\method{as_flextable}{glmmTMB}(x, ...)

\method{as_flextable}{glmmadmb}(x, ...)
}
\arguments{
\item{x}{a mixed model}

\item{...}{unused argument}
}
\description{
produce a flextable describing a
mixed model. The function is only using package 'broom.mixed'
that provides the data presented in the resulting flextable.
}
\examples{
if(require("broom.mixed") && require("nlme")){
  m1 <- lme(distance ~ age, data = Orthodont)
  ft <- as_flextable(m1)
  ft
}
}
