% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.pam}
\alias{as_flextable.pam}
\title{Transform a 'pam' object into a flextable}
\usage{
\method{as_flextable}{pam}(x, digits = 4, ...)
}
\arguments{
\item{x}{a \code{\link[cluster:pam]{cluster::pam()}} object}

\item{digits}{number of digits for the numeric columns}

\item{...}{unused argument}
}
\description{
produce a flextable describing a
pam object. The function is only using package 'broom'
that provides the data presented in the resulting flextable.
}
\examples{
if (require("cluster")) {
  dat <- as.data.frame(scale(mtcars[1:7]))
  cl <- pam(dat, 3)
  ft <- as_flextable(cl)
  ft
}
}
\seealso{
Other as_flextable methods:
\code{\link[=as_flextable.compact_summary]{as_flextable.compact_summary()}},
\code{\link[=as_flextable.data.frame]{as_flextable.data.frame()}},
\code{\link[=as_flextable.gam]{as_flextable.gam()}},
\code{\link[=as_flextable.glm]{as_flextable.glm()}},
\code{\link[=as_flextable.grouped_data]{as_flextable.grouped_data()}},
\code{\link[=as_flextable.htest]{as_flextable.htest()}},
\code{\link[=as_flextable.kmeans]{as_flextable.kmeans()}},
\code{\link[=as_flextable.lm]{as_flextable.lm()}},
\code{\link[=as_flextable.merMod]{as_flextable.merMod()}},
\code{\link[=as_flextable.summarizor]{as_flextable.summarizor()}},
\code{\link[=as_flextable.table]{as_flextable.table()}},
\code{\link[=as_flextable.tabular]{as_flextable.tabular()}},
\code{\link[=as_flextable.tabulator]{as_flextable.tabulator()}},
\code{\link[=as_flextable.xtable]{as_flextable.xtable()}},
\code{\link[=compact_summary]{compact_summary()}}
}
\concept{as_flextable methods}
