% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable}
\alias{as_flextable}
\title{Method to transform objects into flextables}
\usage{
as_flextable(x, ...)
}
\arguments{
\item{x}{object to be transformed as flextable}

\item{...}{arguments for custom methods}
}
\description{
This is a convenient function
to let users create flextable bindings
from any objects. Users should consult documentation
of corresponding method to understand the details and
see what arguments can be used.
}
\seealso{
Other flextable constructors:
\code{\link{flextable}()}
}
\concept{flextable_constructors}
