% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_printer.R
\name{as_flextable.table}
\alias{as_flextable.table}
\title{Transform a 'table' object into a flextable}
\usage{
\method{as_flextable}{table}(x, ...)
}
\arguments{
\item{x}{table object}

\item{...}{arguments used by \code{\link[=proc_freq]{proc_freq()}}.}
}
\description{
produce a flextable describing a
count table produced by function \code{table()}.
}
\examples{
tab <- with(warpbreaks, table(wool, tension))
ft <- as_flextable(tab)
ft
}
\seealso{
Other as_flextable methods: 
\code{\link{as_flextable.data.frame}()},
\code{\link{as_flextable.gam}()},
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.kmeans}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.merMod}()},
\code{\link{as_flextable.pam}()},
\code{\link{as_flextable.summarizor}()},
\code{\link{as_flextable.tabular}()},
\code{\link{as_flextable.tabulator}()},
\code{\link{as_flextable.xtable}()},
\code{\link{as_flextable}()}
}
\concept{as_flextable methods}
