% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.tabular.R
\name{as_flextable.tabular}
\alias{as_flextable.tabular}
\title{convert tabular to flextable}
\usage{
\method{as_flextable}{tabular}(
  x,
  spread_first_col = FALSE,
  fp_p = fp_par(text.align = "center", padding.top = 4),
  row_title = as_paragraph(as_chunk(.row_title)),
  ...
)
}
\arguments{
\item{x}{object produced by \code{\link[tables:tabular]{tables::tabular()}}.}

\item{spread_first_col}{if TRUE, first row is spread as a new line separator
instead of being a column. This helps to reduce the width and allows for
clear divisions.}

\item{fp_p}{paragraph formatting properties associated with row titles,
see \code{\link[=fp_par]{fp_par()}}.}

\item{row_title}{a call to \code{\link[=as_paragraph]{as_paragraph()}} - it
will be applied to the row titles if any
when \code{spread_first_col=TRUE}.}

\item{...}{unused argument}
}
\description{
Produce a flextable from a 'tabular' object
produced with function \code{\link[tables:tabular]{tables::tabular()}}.

When \code{as_flextable.tabular=TRUE}, the first column is
used as row separator acting as a row title. It can
be formated with arguments \code{fp_p} (the formatting
properties of the paragraph) and \code{row_title} that
specifies the content and eventually formattings
of the content.

Two hidden columns can be used after the creation of
the flextable (use only when \code{spread_first_col=TRUE}):
\itemize{
\item \code{.row_title} that contains the title label
\item \code{.is_row_title} that contains TRUE if a row is
considered as a row title.
}
}
\examples{
if (require("tables")) {
  set.seed(42)
  genders <- c("Male", "Female")
  status <- c("low", "medium", "high")
  Sex <- factor(sample(genders, 100, rep = TRUE))
  Status <- factor(sample(status, 100, rep = TRUE))
  z <- rnorm(100) + 5
  fmt <- function(x) {
    s <- format(x, digits = 2)
    even <- ((1:length(s)) \%\% 2) == 0
    s[even] <- sprintf("(\%s)", s[even])
    s
  }
  tab <- tabular(
    Justify(c) * Heading() * z *
      Sex * Heading(Statistic) *
      Format(fmt()) *
      (mean + sd) ~ Status
  )
  as_flextable(tab)

  tab <- tabular(
    (Species + 1) ~ (n = 1) + Format(digits = 2) *
      (Sepal.Length + Sepal.Width) * (mean + sd),
    data = iris
  )
  as_flextable(tab)

  x <- tabular((Factor(gear, "Gears") + 1)
  * ((n = 1) + Percent()
      + (RowPct = Percent("row"))
      + (ColPct = Percent("col")))
  ~ (Factor(carb, "Carburetors") + 1)
    * Format(digits = 1), data = mtcars)

  ft <- as_flextable(
    x, spread_first_col = TRUE,
    row_title = as_paragraph(
      colorize("Gears: ", color = "#666666"),
      colorize(as_b(.row_title), color = "red"))
  )
  ft

  tab <- tabular(
    (mean + mean) * (Sepal.Length + Sepal.Width) ~ 1,
    data = iris
  )
  as_flextable(tab)
}
}
\seealso{
Other as_flextable methods: 
\code{\link{as_flextable.data.frame}()},
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.kmeans}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.merMod}()},
\code{\link{as_flextable.pam}()},
\code{\link{as_flextable.summarizor}()},
\code{\link{as_flextable.tabulator}()},
\code{\link{as_flextable.xtable}()},
\code{\link{as_flextable}()}
}
\concept{as_flextable methods}
