% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable_tabulator.R
\name{as_flextable.tabulator}
\alias{as_flextable.tabulator}
\title{tabulator to flextable}
\usage{
\method{as_flextable}{tabulator}(
  x,
  big_border = fp_border_default(width = 1.5),
  small_border = fp_border_default(width = 0.75),
  sep_w = 0.05,
  ...
)
}
\arguments{
\item{x}{result from \code{\link[=tabulator]{tabulator()}}}

\item{big_border, small_border}{big and small border properties defined
by a call to \code{\link[=fp_border_default]{fp_border_default()}} or \code{\link[=fp_border]{fp_border()}}.}

\item{sep_w}{blank column separators' width to be used. If 0,
blank column separators will not be used.}

\item{...}{unused argument}
}
\description{
\code{tabulator} object can be transformed as a flextable
with method \code{\link[=as_flextable]{as_flextable()}}.
}
\seealso{
\code{\link[=summarizor]{summarizor()}}, \code{\link[=as_grouped_data]{as_grouped_data()}}

Other as_flextable methods: 
\code{\link{as_flextable.gam}()},
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.xtable}()},
\code{\link{as_flextable}()}
}
\concept{as_flextable methods}
