% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_content.R
\name{as_highlight}
\alias{as_highlight}
\title{highlight chunk}
\usage{
as_highlight(x, color)
}
\arguments{
\item{x}{value, if a chunk, the chunk will be updated}

\item{color}{color to use as text highlighting color as character vector.}
}
\description{
The function is producing a chunk with
an highlight chunk.
}
\note{
This is a sugar function that ease the composition of complex
labels made of different formattings. It should be used inside a
call to \code{\link[=as_paragraph]{as_paragraph()}}.
}
\examples{
ft <- flextable( head(iris),
  col_keys = c("Sepal.Length", "dummy") )

ft <- compose(ft, j = "dummy",
  value = as_paragraph(as_highlight(Sepal.Length, color = "yellow")) )

ft
}
\seealso{
Other chunk elements for paragraph: 
\code{\link{as_bracket}()},
\code{\link{as_b}()},
\code{\link{as_chunk}()},
\code{\link{as_image}()},
\code{\link{as_i}()},
\code{\link{as_sub}()},
\code{\link{as_sup}()},
\code{\link{colorize}()},
\code{\link{gg_chunk}()},
\code{\link{hyperlink_text}()},
\code{\link{linerange}()},
\code{\link{lollipop}()},
\code{\link{minibar}()},
\code{\link{plot_chunk}()}
}
\concept{chunk elements for paragraph}
