% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_image}
\alias{as_image}
\title{image}
\usage{
as_image(x, src, width = 1, height = 0.2)
}
\arguments{
\item{x}{dummy variable}

\item{src}{image filename}

\item{width, height}{size of the resulting png file in inches}
}
\description{
image
}
