% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-defunct.R
\name{as_raster}
\alias{as_raster}
\title{Transform a flextable into a raster}
\usage{
as_raster(...)
}
\arguments{
\item{...}{unused}
}
\description{
Function is now defunct, use \code{\link[=gen_grob]{gen_grob()}} instead.
}
\keyword{internal}
