% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_content.R
\name{as_sup}
\alias{as_sup}
\title{Superscript chunk}
\usage{
as_sup(x)
}
\arguments{
\item{x}{value, if a chunk, the chunk will be updated}
}
\description{
The function is producing a chunk with
superscript vertical alignment.

It is used to add it to the content of a cell of the
flextable with the functions \code{\link[=compose]{compose()}}, \code{\link[=append_chunks]{append_chunks()}}
or \code{\link[=prepend_chunks]{prepend_chunks()}}.
}
\note{
This is a sugar function that ease the composition of complex
labels made of different formattings. It should be used inside a
call to \code{\link[=as_paragraph]{as_paragraph()}}.
}
\examples{
ft <- flextable(head(iris), col_keys = c("dummy"))

ft <- compose(ft,
  i = 1, j = "dummy", part = "header",
  value = as_paragraph(
    " anything ",
    as_sup("Sepal.Width")
  )
)

ft <- autofit(ft)
ft
}
\seealso{
Other chunk elements for paragraph: 
\code{\link{as_b}()},
\code{\link{as_bracket}()},
\code{\link{as_chunk}()},
\code{\link{as_equation}()},
\code{\link{as_highlight}()},
\code{\link{as_i}()},
\code{\link{as_image}()},
\code{\link{as_strike}()},
\code{\link{as_sub}()},
\code{\link{as_word_field}()},
\code{\link{colorize}()},
\code{\link{gg_chunk}()},
\code{\link{grid_chunk}()},
\code{\link{hyperlink_text}()},
\code{\link{linerange}()},
\code{\link{minibar}()},
\code{\link{plot_chunk}()}
}
\concept{chunk elements for paragraph}
