% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_flextable.R
\name{body_replace_flextable_at_bkm}
\alias{body_replace_flextable_at_bkm}
\title{Add flextable at bookmark location in a Word document}
\usage{
body_replace_flextable_at_bkm(
  x,
  bookmark,
  value,
  align = "center",
  split = FALSE
)
}
\arguments{
\item{x}{an rdocx object}

\item{bookmark}{bookmark id}

\item{value}{\code{flextable} object}

\item{align}{left, center (default) or right.}

\item{split}{set to TRUE if you want to activate Word
option 'Allow row to break across pages'.}
}
\description{
Use this function if you want to replace a paragraph containing
a bookmark with a flextable. As a side effect, the bookmark will be lost.
}
