% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders.R
\name{border_inner}
\alias{border_inner}
\title{Set all inner borders}
\usage{
border_inner(x, border = NULL, part = "all")
}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}

\item{border}{border properties defined by a call to \code{\link[officer:fp_border]{officer::fp_border()}}}

\item{part}{part selector, see section \emph{Part selection with the \code{part}
parameter} in <\code{\link[=flextable_selectors]{Selectors in flextable}}>.
Value 'all' can be used.}
}
\description{
The function is applying a vertical and horizontal borders to inner content of one
or all parts of a flextable.
}
\examples{
library(officer)
std_border <- fp_border(color = "orange", width = 1)

dat <- iris[c(1:5, 51:55, 101:105), ]
ft <- flextable(dat)
ft <- border_remove(x = ft)

# add inner vertical borders
ft <- border_inner(ft, border = std_border)
ft
}
\seealso{
Other borders management:
\code{\link[=border_inner_h]{border_inner_h()}},
\code{\link[=border_inner_v]{border_inner_v()}},
\code{\link[=border_outer]{border_outer()}},
\code{\link[=border_remove]{border_remove()}},
\code{\link[=hline]{hline()}},
\code{\link[=hline_bottom]{hline_bottom()}},
\code{\link[=hline_top]{hline_top()}},
\code{\link[=surround]{surround()}},
\code{\link[=vline]{vline()}},
\code{\link[=vline_left]{vline_left()}},
\code{\link[=vline_right]{vline_right()}}
}
\concept{borders management}
