% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_grobs.R
\name{dim.flextableGrob}
\alias{dim.flextableGrob}
\title{Get optimal width and height of a flextable grob}
\usage{
\method{dim}{flextableGrob}(x)
}
\arguments{
\item{x}{a flextableGrob object}
}
\value{
a named list with two elements, \code{width} and \code{height}.
Values are expressed in inches.
}
\description{
returns the optimal width and height for the grob,
according to the grob generation parameters.
}
\examples{
ftab <- flextable(head(iris))
gr <- gen_grob(ftab)
dim(gr)
}
