% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim_pretty}
\alias{dim_pretty}
\title{Calculate pretty dimensions}
\usage{
dim_pretty(x)
}
\arguments{
\item{x}{flextable object}
}
\description{
return minimum estimated widths and heights for
each table columns and rows.
}
\examples{

ft <- flextable(mtcars)
\donttest{dim_pretty(ft)}
}
