% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{flextable-data-transform}
\alias{flextable-data-transform}
\alias{as_grouped_data}
\title{data transformations for flextable}
\usage{
as_grouped_data(x, groups, columns = NULL)
}
\arguments{
\item{x}{dataset}

\item{groups}{columns names to be used as row separators.}

\item{columns}{columns names to keep}
}
\description{
Create specific data.frame.
}
\examples{
# as_grouped_data -----
library(data.table)
CO2 <- CO2
setDT(CO2)
CO2$conc <- as.integer(CO2$conc)

data_co2 <- dcast(CO2, Treatment + conc ~ Type,
  value.var = "uptake", fun.aggregate = mean)
data_co2
data_co2 <- as_grouped_data(x = data_co2, groups = c("Treatment"))
data_co2

}
\seealso{
\code{\link{as_flextable}}
}
