% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{flextable-defunct}
\alias{flextable-defunct}
\alias{ph_with_flextable_at}
\alias{display}
\title{Defunct Functions in Package flextable}
\usage{
ph_with_flextable_at(...)

display(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package flextable
}
\details{
\code{ph_with_flextable_at()} is replaced by \code{officer::ph_with}.

\code{display()} is replaced by \code{compose}.
}
