% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_str.R
\name{flextable_html_dependency}
\alias{flextable_html_dependency}
\title{htmlDependency for flextable objects}
\usage{
flextable_html_dependency(htmlscroll = TRUE)
}
\arguments{
\item{htmlscroll}{add a scroll if table is too big to fit
into its HTML container, default to TRUE.}
}
\description{
When using loops in an R Markdown for HTML document, the
htmlDependency object for flextable must also be added at least once.
}
\examples{
if(require("htmltools"))
  div(flextable_html_dependency())
}
