% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{width}
\alias{autofit}
\alias{height}
\alias{width}
\title{widths and heights}
\usage{
width(x, j = NULL, width)

height(x, i = NULL, height, part = "body")

autofit(x, add_w = 0.1, add_h = 0.1)
}
\arguments{
\item{x}{a flextable object}

\item{j}{columns selection}

\item{width}{width in inches}

\item{i}{rows selection}

\item{height}{height in inches}

\item{part}{partname of the table}

\item{add_w}{extra width to add in inches}

\item{add_h}{extra height to add in inches}
}
\description{
modify columns widths with function \code{width}.

modify row heights with function \code{height}.

compute and apply optimized widths and heights with function \code{autofit}.
}
\details{
Heights are not used when flextable is been rendered into HTML.
}
\examples{

ft <- flextable(iris)
ft <- width(ft, width = 1)

write_docx("autofit.docx", ft)

ft <- flextable(iris)
ft <- height(ft, height = .3)

write_docx("height.docx", ft)

ft <- flextable(mtcars)
ft <- autofit(ft)

write_docx("autofit.docx", ft)
}
\seealso{
\code{\link{flextable}}
}

