% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizor.R
\name{fmt_header_n}
\alias{fmt_header_n}
\title{format count data for headers}
\usage{
fmt_header_n(n)
}
\arguments{
\item{n}{count values}
}
\description{
The function formats counts as \verb{\\n(N=XX)}. This helper
function is used to add counts in columns titles.
}
\seealso{
\code{\link[=tabulator]{tabulator()}}, \code{\link[=mk_par]{mk_par()}}

Other other text formatter functions: 
\code{\link{fmt_2stats}()},
\code{\link{fmt_n_percent}()}
}
\concept{other text formatter functions}
