% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-defunct.R
\name{footers_flextable_at_bkm}
\alias{footers_flextable_at_bkm}
\title{Add flextable at a bookmark location in document's footer}
\usage{
footers_flextable_at_bkm(...)
}
\arguments{
\item{...}{unused}
}
\description{
Function is now defunct, use \code{\link[officer:prop_section]{officer::prop_section()}} instead.
}
\keyword{internal}
