% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{format.flextable}
\alias{format.flextable}
\title{Encode flextable in a document format.}
\usage{
\method{format}{flextable}(x, type, ...)
}
\arguments{
\item{x}{flextable object}

\item{type}{one of pptx, docx or html.}

\item{...}{unused}
}
\description{
Encode flextable in a document format, \code{html}, \code{docx},
\code{pptx}.

This function is exported so that users can create their own custom
component.
}
\examples{
ft <- flextable(head(iris, n = 2))
format(ft, type = "html")
}
\seealso{
Other flextable print function: 
\code{\link{as_raster}()},
\code{\link{docx_value}()},
\code{\link{htmltools_value}()},
\code{\link{knit_print.flextable}()},
\code{\link{plot.flextable}()},
\code{\link{print.flextable}()},
\code{\link{save_as_html}()},
\code{\link{save_as_image}()}
}
\concept{flextable print function}
