% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_that.R
\name{format_simple}
\alias{format_simple}
\title{format_simple}
\usage{
format_simple(expr, pr_text_ = fp_text(), pr_par_ = fp_par())
}
\arguments{
\item{expr}{a call to \code{ftext} or \code{external_img}}

\item{pr_text_}{default text properties}

\item{pr_par_}{default paragraph properties}
}
\description{
format_simple generic function
}
\examples{
format_simple("hello", pr_text_ = fp_text())
}

