% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizor.R
\name{format_summarizor_cols}
\alias{format_summarizor_cols}
\title{format content for data generated with \code{\link[=summarizor]{summarizor()}}}
\usage{
format_summarizor_cols(stat, data_type, value, count, percent)
}
\arguments{
\item{stat, data_type, value, count, percent}{column names
generated by summarizor().}
}
\description{
format content for data generated with \code{\link[=summarizor]{summarizor()}}
}
