% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_that.R
\name{format_that}
\alias{format_that}
\title{format_that}
\usage{
format_that(str, pr_text_ = pr_text(), pr_par_ = pr_par(), ...)
}
\arguments{
\item{str}{string template}

\item{pr_text_}{default text properties}

\item{pr_par_}{default paragraph properties}

\item{...}{named arguments, element should be a call to \code{ftext} or \code{external_img}}
}
\description{
format_that generic function.
}
\examples{
format_that("I like to {{ x }} it", x = ftext("move", t_italic()), prop = pr_text())
}

