% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_functions.R
\name{format_pvalue}
\alias{format_default}
\alias{format_double}
\alias{format_integer}
\alias{format_pvalue}
\alias{formatting_functions}
\title{formatting functions}
\usage{
format_pvalue(x)

format_double(x, digits = 3)

format_integer(x)

format_default(x)
}
\arguments{
\item{x}{the vector to format}

\item{digits}{number of digits after the decimal point}
}
\description{
formatting functions to use with lsmeans objects
}
\section{extends}{

Pour creer une nouvelle fonction de formatage, il faut une fonction
qui ne prend qu'un argument (un vecteur) et qui renvoie une chaine de
caracteres de meme taille que le vecteur d'entree.
}
\examples{
format_pvalue( runif(10))

format_percent <- function(x)
  ifelse( is.na(x), "", paste0(formatC(x*100, format = "f", digits = 1 ), "\%" ) )
format_percent( runif(10))
format_double( runif(10))
format_integer( rnorm(10, 3))
format_integer( rnorm(10, 3))
}

