% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paragraph.R
\name{ftext}
\alias{dim.ftext}
\alias{format.ftext}
\alias{ftext}
\title{formatted text}
\usage{
ftext(text, prop)

\method{format}{ftext}(x, type = "wml", ...)

\method{dim}{ftext}(x)
}
\arguments{
\item{text}{text value}

\item{prop}{formatting text properties}

\item{x}{\code{ftext} object}

\item{type}{output format}

\item{...}{unused}
}
\description{
format a chunk of text with a text properties attribute
}
\section{dim}{


Function \code{dim} returns width and height of the \code{ftext} when
formatted.
}
\examples{
ftext("hello", fp_text())
}

