% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_flextable_defaults}
\alias{get_flextable_defaults}
\title{Get flextable defaults formatting properties}
\usage{
get_flextable_defaults()
}
\value{
a list containing default values.
}
\description{
The current formatting properties
are automatically applied to every flextable you produce.
These default values are returned by this function.
}
\examples{
get_flextable_defaults()
}
\seealso{
Other functions related to themes: 
\code{\link{set_flextable_defaults}()},
\code{\link{theme_alafoli}()},
\code{\link{theme_apa}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_borderless}()},
\code{\link{theme_box}()},
\code{\link{theme_tron}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{functions related to themes}
