% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{get_graphic_frame}
\alias{get_graphic_frame}
\title{graphic frame xml code}
\usage{
get_graphic_frame(x, id = 1L, offx = 1L, offy = 1L)
}
\arguments{
\item{x}{\code{flextable} object}

\item{id}{unique identifier in the slide}

\item{offx}{x offset}

\item{offy}{y offset}

\item{size}{a named vector containing width and height in inches}
}
\description{
open xml wrapper for pptx
}

