% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{header_col}
\alias{header_col}
\title{Set the column that contains row headers}
\usage{
header_col(x, j = 1, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{j}{column selection}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
set the column which contains row headers. Only applies to html tables.
}
\examples{
ft <- flextable(mtcars)
ft <- header_col(ft, j = 1)
}
\seealso{
Other sugar functions for table style: 
\code{\link{align}()},
\code{\link{bg}()},
\code{\link{bold}()},
\code{\link{color}()},
\code{\link{empty_blanks}()},
\code{\link{fontsize}()},
\code{\link{font}()},
\code{\link{italic}()},
\code{\link{keepnext}()},
\code{\link{padding}()},
\code{\link{rotate}()},
\code{\link{valign}()}
}
\concept{sugar functions for table style}
