% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{height}
\alias{height}
\title{set columns width}
\usage{
height(x, i = NULL, height, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{height}{height in inches}

\item{part}{partname of the table}
}
\description{
modify columns widths of a flextable
}
\examples{

ft <- vanilla_table(iris)
ft <- height(ft, height = .3)

write_docx("height.docx", ft)
}
\seealso{
\code{\link{flextable}}
}

