% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_content.R
\name{i}
\alias{i}
\title{bold chunk}
\usage{
i(x)
}
\arguments{
\item{x}{text or any element that can be formatted as italic}
}
\description{
The function is producing a chunk with
italic font. This is a sugar function
written mainly to ease the composition of complex header labels.
It should be used inside a call to \code{\link{as_paragraph}}.
}
\seealso{
Other chunk elements for paragraph: \code{\link{as_chunk}},
  \code{\link{as_image}}, \code{\link{b}},
  \code{\link{hyperlink_text}}, \code{\link{linerange}},
  \code{\link{minibar}}, \code{\link{sub}},
  \code{\link{sup}}
}
\concept{chunk elements for paragraph}
