% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{information_data_paragraph}
\alias{information_data_paragraph}
\title{paragraph related information of a flextable}
\usage{
information_data_paragraph(x)
}
\arguments{
\item{x}{a flextable object}
}
\value{
a data.frame containing information about paragraphs:
\itemize{
\item formatting properties,
\item part (\code{.part}), row (\code{.row_id}) and column (\code{.col_id}).
}
}
\description{
This function takes a flextable object and returns a data.frame containing
information about each paragraph within the flextable. The data.frame includes
details about formatting properties and position within the
row and column.
}
\section{don't use this}{


These data structures should not be used, as they
represent an interpretation of the underlying data
structures, which may evolve over time.

\strong{They are exported to enable two packages that exploit
these structures to make a transition, and should not
remain available for long.}
}

\examples{
ft <- as_flextable(iris)
x <- information_data_paragraph(ft)
head(x)
}
\seealso{
Other information data functions: 
\code{\link{information_data_cell}()},
\code{\link{information_data_chunk}()}
}
\concept{information data functions}
\keyword{internal}
