% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{keepnext}
\alias{keepnext}
\title{Set Keep with next attribute}
\usage{
keepnext(x, i = NULL, keepnext = TRUE, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{keepnext}{TRUE to set Keep with next, FALSE to unset}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
set or unset "Keep with next" attribute of selected rows of a flextable. Only applies when inserting into a Word document.
}
\examples{
ft <- flextable(mtcars)
ft <- keepnext(ft)
}
\seealso{
Other sugar functions for table style: 
\code{\link{align}()},
\code{\link{bg}()},
\code{\link{bold}()},
\code{\link{color}()},
\code{\link{empty_blanks}()},
\code{\link{fontsize}()},
\code{\link{font}()},
\code{\link{italic}()},
\code{\link{padding}()},
\code{\link{rotate}()},
\code{\link{valign}()}
}
\concept{sugar functions for table style}
