% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{labelize}
\alias{labelize}
\title{change displayed labels}
\usage{
labelize(x, j, labels, part = "all")
}
\arguments{
\item{x}{a flextable object}

\item{j}{columns selection}

\item{labels}{a named vector whose names will be used to identify
values to replace and values will be used as labels.}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
The function replace text values
in a flextable with labels.
}
\examples{
z <- summarizor(
  x = CO2[-c(1, 4)],
  by = "Treatment",
  overall_label = "Overall")

tab_1 <- tabulator(
  x = z,
  rows = c("variable", "stat"),
  columns = "Treatment",
  blah = as_paragraph(
    as_chunk(
      fmt_2stats(
        stat = stat,
        num1 = value1, num2 = value2,
        cts = cts, pcts = percent
      )
    )
  )
)

ft_1 <- as_flextable(tab_1, separate_with = "variable")


ft_1 <- labelize(
  x = ft_1, j = c("stat", "variable"),
  labels = c(n = "N", avg_dev = "Mean (SD)", median = "Median",
             range = "Range", missing = "Missing",
             stat = "Hello")
)

ft_1
}
