% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{latex_value}
\alias{latex_value}
\title{flextable latex string for PDF}
\usage{
latex_value(
  x,
  print = TRUE,
  ft.align = opts_current$get("ft.align"),
  ft.tabcolsep = opts_current$get("ft.tabcolsep"),
  ft.arraystretch = opts_current$get("ft.arraystretch"),
  bookdown
)
}
\arguments{
\item{x}{a flextable object}

\item{print}{print output if TRUE}

\item{ft.align}{flextable alignment, supported values are 'left', 'center' and 'right'.}

\item{ft.tabcolsep}{space between the text and the left/right border of its containing
cell to 6 points, the default value is 8 points.}

\item{ft.arraystretch}{height of each row relative to its default
height, the default value is 1.5.}

\item{bookdown}{\code{TRUE} or \code{FALSE} (default) to support cross referencing with bookdown.}
}
\description{
get latex raw code for PDF
from a flextable object.

The function is particularly useful when you want
to generate flextable in a loop from a R Markdown document.
By default, the output is printed and is returned as a
character scalar.

\strong{When used inside an R Markdown document, chunk option \code{results}
must be set to 'asis'.}

Arguments \code{ft.align}, \code{ft.tabcolsep} and \code{ft.arraystretch} can be
specified also as knitr chunk options.
}
\examples{
latex_value(flextable(airquality[1:5,]))
}
\seealso{
Other flextable print function: 
\code{\link{as_raster}()},
\code{\link{docx_value}()},
\code{\link{htmltools_value}()},
\code{\link{knit_print.flextable}()},
\code{\link{plot.flextable}()},
\code{\link{print.flextable}()},
\code{\link{save_as_docx}()},
\code{\link{save_as_html}()},
\code{\link{save_as_image}()},
\code{\link{save_as_pptx}()}
}
\concept{flextable print function}
