% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_flextable.R
\name{merge_v}
\alias{merge_h}
\alias{merge_v}
\title{merge flextable cells}
\usage{
merge_v(x, j = NULL, part = "body")

merge_h(x, i = NULL, part = "body")
}
\arguments{
\item{x}{\code{flextable} object}

\item{j}{columns names/keys where cells have to be merged.}

\item{part}{partname of the table where merge has to be done.}

\item{i}{rows where cells have to be merged.}
}
\description{
Merge flextable cells consecutive identical values.
}
\examples{
ft_merge <- flextable(mtcars)
ft_merge <- merge_v(ft_merge, j = c("gear", "carb"), part = "body")
ft_merge <- merge_h(ft_merge, part = "body")
write_docx("ft_merge.docx", ft_merge)
}

