% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_flextable.R
\name{merge_v}
\alias{merge_v}
\title{Merge flextable cells vertically}
\usage{
merge_v(x, j = NULL, part = "body")
}
\arguments{
\item{x}{\code{flextable} object}

\item{j}{columns names/keys where cells have to be merged.}

\item{part}{partname of the table where merge has to be done.}
}
\description{
Merge flextable cells vertically when consecutive cells have
identical values.
}
\examples{
ft_merge <- flextable(mtcars)
ft_merge <- merge_v(ft_merge, j = c("gear", "carb"))
ft_merge
}
\seealso{
Other flextable merging function: \code{\link{merge_at}},
  \code{\link{merge_h_range}}, \code{\link{merge_h}},
  \code{\link{merge_none}}
}
\concept{flextable merging function}
