% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_themes.R
\name{my_theme_booktabs}
\alias{my_theme_booktabs}
\title{Apply my modified booktabs theme}
\usage{
my_theme_booktabs(x, fontsize = 9, left_just_cols = 1)
}
\arguments{
\item{x}{a flextable object}

\item{fontsize}{font size in pixel}

\item{left_just_cols}{column index where left justify should stop (starts at 1)}
}
\description{
Apply theme booktabs to a flextable
}
\examples{
ftab <- flextable(iris)
ftab <- theme_booktabs(ftab)
}
\seealso{
Other flextable theme: 
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{flextable theme}
