% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrowcol.R
\name{ncol_keys}
\alias{ncol_keys}
\title{Number of columns}
\usage{
ncol_keys(x)
}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}
}
\description{
returns the number of columns displayed
}
\examples{
library(flextable)
ft <- qflextable(head(cars))
ncol_keys(ft)
}
\seealso{
Other functions for flextable size management: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{nrow_part}()},
\code{\link{set_table_properties}()},
\code{\link{width}()}
}
\concept{functions for flextable size management}
