% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_images.R
\name{note_docx_images}
\alias{note_docx_images}
\title{internal utils for roxygen tags reuse}
\description{
internal utils for roxygen tags reuse
}
\note{
This chunk option requires package officedown in a R Markdown
context with Word output format. With Quarto (\code{format: docx}) or
\code{rmarkdown::word_document()}, the resulting file must be repaired
with \code{\link[=repair_docx]{repair_docx()}}.

PowerPoint cannot mix images and text in a paragraph, images
are removed when outputing to PowerPoint format.
}
\keyword{internal}
