% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nrow_part}
\alias{nrow_part}
\title{Number of rows for each part}
\usage{
nrow_part(x, part = "body")
}
\arguments{
\item{x}{flextable object}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
returns the number of lines in each part of the flextable.
}
\examples{
library(flextable)
ft <- qflextable(head(cars))
nrow_part(ft, part = "body")
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{ncol_keys}()},
\code{\link{set_table_properties}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
