% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{padding}
\alias{padding}
\title{Set paragraph paddings}
\usage{
padding(x, i = NULL, j = NULL, padding = NULL, padding.top = NULL,
  padding.bottom = NULL, padding.left = NULL, padding.right = NULL,
  part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{padding}{padding (shortcut for top, bottom, left and right)}

\item{padding.top}{padding top}

\item{padding.bottom}{padding bottom}

\item{padding.left}{padding left}

\item{padding.right}{padding right}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
change paddings of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- padding(ft, padding.top = 4)
}
\seealso{
Other sugar functions for table style: \code{\link{align}},
  \code{\link{bg}}, \code{\link{bold}},
  \code{\link{color}}, \code{\link{empty_blanks}},
  \code{\link{fontsize}}, \code{\link{font}},
  \code{\link{italic}}, \code{\link{rotate}},
  \code{\link{valign}}
}
\concept{sugar functions for table style}
