% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_grobs.R
\name{plot.flextableGrob}
\alias{plot.flextableGrob}
\title{plot a flextable grob}
\usage{
\method{plot}{flextableGrob}(x, ...)
}
\arguments{
\item{x}{a flextableGrob object}

\item{...}{additional arguments passed to other functions}
}
\description{
plot a flextable grob
}
