% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{pptx_add_flextable}
\alias{pptx_add_flextable}
\title{wml table code}
\usage{
pptx_add_flextable(x, value, id, index = 1)
}
\arguments{
\item{x}{a pptx device}

\item{value}{\code{flextable} object}

\item{id}{placeholder id}

\item{index}{placeholder index. This is to be used when a placeholder id
is not unique in the current slide, e.g. two placeholders with id 'body'.}
}
\description{
produces the wml of a flextable
}

