% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_freq.R
\name{proc_freq}
\alias{proc_freq}
\title{frequency table as flextable}
\usage{
proc_freq(x, row, col, main = "")
}
\arguments{
\item{x}{\code{data.frame} object}

\item{row}{\code{characer} column names for row}

\item{col}{\code{characer} column names for column}

\item{main}{\code{characer} title}
}
\description{
This function compute a two way contingency table
and make a flextable with the result.
}
\examples{

proc_freq(mtcars, "vs", "gear")
proc_freq(mtcars, "gear", "vs")
proc_freq(mtcars, "gear", "vs", "My title")
}
\author{
Titouan Robert
}
