% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procfreq.R
\name{procFreq}
\alias{procFreq}
\title{procFreq for flextable}
\usage{
procFreq(x, row, col)
}
\arguments{
\item{x}{\code{data.frame} object
value specifying label to use.}

\item{row}{\code{characer} column of x for row}

\item{col}{\code{characer} column of x for col}
}
\description{
This function compute statistics and make a flextable.
}
\examples{
data("mtcars")
procFreq(mtcars, "vs", "gear")
procFreq(mtcars, "gear", "vs")

}
