% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R, R/rtf_str.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{to_html}
\alias{to_rtf}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{officer}{\code{\link[officer]{to_html}}, \code{\link[officer]{to_rtf}}}
}}

