% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-deprecated.R
\name{regulartable}
\alias{regulartable}
\title{flextable old functions}
\usage{
regulartable(data, col_keys = names(data), cwidth = 0.75, cheight = 0.25)
}
\arguments{
\item{data}{dataset}

\item{col_keys}{columns names/keys to display. If some column names are not in
the dataset, they will be added as blank columns by default.}

\item{cwidth, cheight}{initial width and height to use for cell sizes in inches.}
}
\description{
The function is maintained for compatibility with old codes
mades by users but be aware it produces the same exact object than \code{\link[=flextable]{flextable()}}.
This function should be deprecated then removed in the next versions.
}
\keyword{internal}
