% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_rmarkdown.R
\name{render_flextable}
\alias{render_flextable}
\title{Render flextable in rmarkdown (including Word output)}
\usage{
render_flextable(x, ...)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{...}{further arguments}
}
\description{
Function to use in the knitr/rmarkdown's \code{render} chunk 
option.
For Word (docx) output, if pandoc vesion >= 2.0 is used, a raw XML block 
with the table code will be inserted.
For HTML output, you won't need an extra call to \code{\link{tabwid}}, but 
can just use \code{\link{flextable}} inside a chunk.
}
\note{
To insert all flextables automatically, define 
\code{knit_print.flextable = render_flextable} in the beginning of 
the rmarkdown document.
}
\seealso{
vignette("knit_print", package = "knitr")
}
\author{
Maxim Nazarov
}
