% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{save_as_image}
\alias{save_as_image}
\title{save a flextable as an image}
\usage{
save_as_image(x, file, zoom = 3, expand = 10)
}
\arguments{
\item{x}{a flextable object}

\item{zoom, expand}{parameters used by \code{webshot} function.}

\item{path}{image file to be created. It should end with .png, .pdf, or .jpeg.}
}
\description{
save a flextable as an image. This function
require webshot package.
}
\examples{
ft <- flextable( head( mtcars ) )
ft <- autofit(ft)
tf <- tempfile(fileext = ".png")
save_as_image(ft, tf)
}
\seealso{
Other flextable print function: \code{\link{format.flextable}},
  \code{\link{htmltools_value}},
  \code{\link{knit_print.flextable}},
  \code{\link{print.flextable}}, \code{\link{save_as_html}}
}
\concept{flextable print function}
