% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set_caption}
\alias{set_caption}
\title{set caption}
\usage{
set_caption(x, caption, html_escape = TRUE)
}
\arguments{
\item{x}{flextable object}

\item{caption}{caption value}

\item{html_escape}{should HTML entities be escaped so that it can be safely
included as text or an attribute value within an HTML document.}
}
\description{
set caption value in flextable
}
\note{
this will have an effect only when output is HTML or Word document.
}
\examples{
ft <- flextable( head( iris ) )
ft <- set_caption(ft, "my caption")
ft
}
