% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{set_flextable_defaults}
\alias{set_flextable_defaults}
\title{Modify flextable defaults formatting properties}
\usage{
set_flextable_defaults(
  font.family = NULL,
  font.size = NULL,
  font.color = NULL,
  text.align = NULL,
  padding.bottom = NULL,
  padding.top = NULL,
  padding.left = NULL,
  padding.right = NULL,
  border.color = NULL,
  background.color = NULL,
  table.layout = NULL,
  theme_fun = NULL
)
}
\arguments{
\item{font.family}{single character value specifying font name.}

\item{font.size}{font size (in point) - 0 or positive integer value.}

\item{font.color}{font color - a single character value specifying
a valid color (e.g. "#000000" or "black").}

\item{text.align}{text alignment - a single character value, expected value
is one of 'left', 'right', 'center', 'justify'.}

\item{padding.bottom, padding.top, padding.left, padding.right}{paragraph paddings - 0 or
positive integer value.}

\item{border.color}{border color - single character value
(e.g. "#000000" or "black").}

\item{background.color}{cell background color - a single character value specifying a
valid color (e.g. "#000000" or "black").}

\item{table.layout}{'autofit' or 'fixed' algorithm. Default to 'autofit'.}

\item{theme_fun}{a single character value, the name of the theme function
to be applied}
}
\value{
a list containing previous default values.
}
\description{
The current formatting properties (see \code{\link[=get_flextable_defaults]{get_flextable_defaults()}})
are automatically applied to every flextable you produce.
Use \code{set_flextable_defaults()} to override them.
}
\examples{
ft_1 <- flextable(head(mtcars))
ft_1
set_flextable_defaults(
  table.layout = "autofit",
  font.size = 12,  font.color = "gray10",
  background.color = "gray90")

set_flextable_defaults(theme_fun = "theme_vanilla" )
ft_2 <- flextable(head(mtcars))
ft_2

set_flextable_defaults(theme_fun = "theme_vader" )
ft_3 <- flextable(head(mtcars))
ft_3
}
\seealso{
Other functions related to themes: 
\code{\link{get_flextable_defaults}()},
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{functions related to themes}
