% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz-defunct.R
\name{set_formatter_type}
\alias{set_formatter_type}
\title{Set Formatter by Types of Columns}
\usage{
set_formatter_type(...)
}
\description{
Function is now defunct, use \code{colformat_} functions instead.
}
\keyword{internal}
