% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set_header}
\alias{set_header}
\title{set flextable's headers}
\usage{
set_header(x, ..., data_mapping = NULL, key = "col_keys", values = NULL)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{...}{a named list (names are data colnames) with character
values specifying in reverse order content of the column.}

\item{data_mapping}{a \code{data.frame} specyfing for each colname
content of the column.}

\item{key}{column to use as key when joigning data_mapping.}

\item{values}{character vector indicating colnames' labels.}
}
\description{
The definition of flextable's headers can be done
in different manner.
}
\examples{

# set_header - method 1 ------
ft_1 <- flextable( head( iris ))
ft_1 <- set_header(ft_1,
  values = c("Sepal length", "Sepal width",
             "Petal length", "Petal width", "Species") )
write_docx("ft_1.docx", ft_1)

# set_header - method 2 ------
ft_2 <- flextable( head( iris ))
ft_2 <- set_header(x = ft_2,
  Sepal.Length = list("Sepal", "Length"),
  Sepal.Width = list("Sepal", "Width"),
  Petal.Length = list("Petal", "Length"),
  Petal.Width = list("Petal", "Width"),
  Species = list("Species", "Species") )
write_docx("ft_2.docx", ft_2)

# set_header - method 3 ------
typology <- data.frame(
  col_keys = c( "Sepal.Length", "Sepal.Width", "Petal.Length",
                "Petal.Width", "Species" ),
  what = c("Sepal", "Sepal", "Petal", "Petal", "Species"),
  measure = c("Length", "Width", "Length", "Width", "Species"),
  stringsAsFactors = FALSE )

ft_3 <- flextable( head( iris ))
ft_3 <- set_header(ft_3, data_mapping = typology, key = "col_keys" )
write_docx("ft_3.docx", ft_3)
}

