% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set_header_df}
\alias{set_header_df}
\title{set flextable's header rows}
\usage{
set_header_df(x, mapping = NULL, key = "col_keys")
}
\arguments{
\item{x}{a \code{flextable} object}

\item{mapping}{a \code{data.frame} specyfing for each colname
content of the column.}

\item{key}{column to use as key when joigning data_mapping.}
}
\description{
Use a data.frame to specify flextable's header rows.
}
\examples{
typology <- data.frame(
  col_keys = c( "Sepal.Length", "Sepal.Width", "Petal.Length",
                "Petal.Width", "Species" ),
  what = c("Sepal", "Sepal", "Petal", "Petal", "Species"),
  measure = c("Length", "Width", "Length", "Width", "Species"),
  stringsAsFactors = FALSE )

ft <- flextable( head( iris ))
ft <- set_header_df(ft, mapping = typology, key = "col_keys" )
write_docx("header_df.docx", ft)
}

