% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{set_table_properties}
\alias{set_table_properties}
\title{Global table properties}
\usage{
set_table_properties(x, layout = "fixed", width = 1)
}
\arguments{
\item{x}{flextable object}

\item{layout}{'autofit' or 'fixed' algorithm. Default to 'autofit'.}

\item{width}{value of the preferred width of the table in percent.}
}
\description{
Set table layout and table width. Default to fixed
algorithm.

If layout is fixed, column widths will be used to display the table;
\code{width} is ignored.

If layout is autofit, column widths will not be used;
table width is used (as a percentage).
}
\note{
PowerPoint output ignore autofit layout as this algorithm does not
exist for this Microsoft format.
}
\examples{
library(flextable)
ft_1 <- flextable(head(mtcars))
ft_1 <- set_table_properties(ft_1, width = .8, layout = "autofit")
print(ft_1)
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
