% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style}
\alias{style}
\title{Set flextable default styles}
\usage{
style(
  x,
  i = NULL,
  j = NULL,
  pr_t = NULL,
  pr_p = NULL,
  pr_c = NULL,
  part = "body"
)
}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}

\item{i}{row selector, see section \emph{Row selection with the \code{i} parameter}
in <\code{\link[=flextable_selectors]{Selectors in flextable}}>.}

\item{j}{column selector, see section \emph{Column selection with the \code{j} parameter}
in <\code{\link[=flextable_selectors]{flextable column selection}}>.}

\item{pr_t}{object(s) of class \code{fp_text}, result of \code{\link[officer:fp_text]{officer::fp_text()}}
or \code{\link[officer:fp_text]{officer::fp_text_lite()}}}

\item{pr_p}{object(s) of class \code{fp_par}, result of \code{\link[officer:fp_par]{officer::fp_par()}}
or \code{\link[officer:fp_par]{officer::fp_par_lite()}}}

\item{pr_c}{object(s) of class \code{fp_cell}, result of \code{\link[officer:fp_cell]{officer::fp_cell()}}}

\item{part}{part selector, see section \emph{Part selection with the \code{part}
parameter} in <\code{\link[=flextable_selectors]{Selectors in flextable}}>.
Value 'all' can be used.}
}
\description{
Modify flextable text, paragraph, and cell default formatting
properties. This allows you to specify a set of formatting properties for a
selection instead of using multiple functions (e.g., \code{bold}, \code{italic}, \code{bg})
that must all be applied to the same selection of rows and columns.
}
\examples{
library(officer)
def_cell <- fp_cell(border = fp_border(color = "wheat"))

def_par <- fp_par(text.align = "center")

ft <- flextable(head(mtcars))

ft <- style(ft, pr_c = def_cell, pr_p = def_par, part = "all")
ft <- style(ft, ~ drat > 3.5, ~ vs + am + gear + carb,
  pr_t = fp_text(color = "red", italic = TRUE)
)

ft
}
