% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_content.R
\name{sup}
\alias{sup}
\title{superscript chunk}
\usage{
sup(x)
}
\arguments{
\item{x}{text or any element that can be formatted as superscript}
}
\description{
The function is producing a chunk with
superscript vertical alignment. This is a sugar function
written mainly to ease the composition of complex header labels.
It should be used inside a call to \code{\link{as_paragraph}}.
}
\seealso{
Other chunk elements for paragraph: \code{\link{as_chunk}},
  \code{\link{as_image}}, \code{\link{b}},
  \code{\link{hyperlink_text}}, \code{\link{i}},
  \code{\link{linerange}}, \code{\link{minibar}},
  \code{\link{sub}}
}
\concept{chunk elements for paragraph}
