% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_str.R
\name{tabwid}
\alias{tabwid}
\title{htmlwidget for flextable}
\usage{
tabwid(x, ...)
}
\arguments{
\item{x}{\code{flextable} object}

\item{...}{Unused arguments}
}
\description{
htmlwidget for flextable. This function is deprecated.
}
