% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_booktabs}
\alias{theme_booktabs}
\title{Apply booktabs theme}
\usage{
theme_booktabs(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme tron to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_booktabs(ft)
}
