% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_vanilla}
\alias{theme_tron}
\alias{theme_tron_legacy}
\alias{theme_vanilla}
\alias{theme_zebra}
\title{flextable theme}
\usage{
theme_vanilla(x)

theme_zebra(x, odd_header = "#CFCFCF", odd_body = "#EFEFEF",
  even_header = "transparent", even_body = "transparent")

theme_tron_legacy(x)

theme_tron(x)
}
\arguments{
\item{x}{a flextable object}

\item{odd_header, odd_body, even_header, even_body}{odd/even colors for table header and body}
}
\description{
apply a theme to a flextable
}
\examples{
# theme_vanilla -------
ft <- flextable(iris)
ft <- theme_tron(ft)
ft
# theme_zebra -------
ft <- flextable(iris)
ft <- theme_zebra(ft)
# theme_tron_legacy -------
ft <- flextable(iris)
ft <- theme_tron_legacy(ft)
# theme_tron -------
ft <- flextable(iris)
ft <- theme_tron(ft)
}
\seealso{
\code{\link{flextable}}
}

