% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tron}
\alias{theme_tron}
\title{Apply tron theme}
\usage{
theme_tron(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme tron to a flextable
}
\examples{
ftab <- flextable(iris)
ftab <- theme_tron(ftab)
}
\seealso{
Other flextable theme: 
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{flextable theme}
