% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tron_legacy}
\alias{theme_tron_legacy}
\title{Apply tron legacy theme}
\usage{
theme_tron_legacy(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme tron legacy to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_tron_legacy(ft)
}
\seealso{
Other flextable theme: 
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{flextable theme}
