% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_vader}
\alias{theme_vader}
\title{Apply Sith Lord Darth Vader}
\usage{
theme_vader(x, ...)
}
\arguments{
\item{x}{a flextable object}

\item{...}{unused}
}
\description{
Apply Sith Lord Darth Vader theme to a flextable
}
\examples{
ftab <- flextable(iris)
ftab <- theme_vader(ftab)
}
\seealso{
Other functions related to themes: 
\code{\link{get_flextable_defaults}()},
\code{\link{set_flextable_defaults}()},
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{functions related to themes}
