% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_vanilla}
\alias{theme_vanilla}
\title{Apply vanilla theme}
\usage{
theme_vanilla(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme vanilla to a flextable
}
\examples{
ftab <- flextable(iris)
ftab <- theme_vanilla(ftab)
}
\seealso{
Other functions related to themes: 
\code{\link{get_flextable_defaults}()},
\code{\link{set_flextable_defaults}()},
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_zebra}()}
}
\concept{functions related to themes}
