% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{to_html.flextable}
\alias{to_html.flextable}
\title{Get HTML code as a string}
\usage{
\method{to_html}{flextable}(x, type = c("table", "img"), ...)
}
\arguments{
\item{x}{a flextable object}

\item{type}{output type. one of "table" or "img".}

\item{...}{unused}
}
\value{
If \code{type='img'}, the result will be a string
containing HTML code of an image tag, otherwise, the
result will be a string containing HTML code of
a table tag.
}
\description{
Generate HTML code of corresponding
flextable as an HTML table or an HTML image.
}
\examples{
library(officer)
library(flextable)
x <- to_html(as_flextable(cars))
}
\seealso{
Other flextable print function: 
\code{\link{df_printer}()},
\code{\link{flextable_to_rmd}()},
\code{\link{gen_grob}()},
\code{\link{htmltools_value}()},
\code{\link{knit_print.flextable}()},
\code{\link{plot.flextable}()},
\code{\link{print.flextable}()},
\code{\link{save_as_docx}()},
\code{\link{save_as_html}()},
\code{\link{save_as_image}()},
\code{\link{save_as_pptx}()},
\code{\link{save_as_rtf}()}
}
\concept{flextable print function}
