% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_printer.R
\name{use_df_printer}
\alias{use_df_printer}
\title{Summarize a data.frame as a flextable}
\usage{
use_df_printer()
}
\description{
Define \code{\link[=df_printer]{df_printer()}} as data.frame
print method in an R Markdown document.

In a setup run chunk:\preformatted{flextable::use_df_printer()
}
}
\seealso{
\code{\link[=df_printer]{df_printer()}}, \code{\link[=flextable]{flextable()}}
}
