% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{use_model_printer}
\alias{use_model_printer}
\title{set model automatic printing as a flextable}
\usage{
use_model_printer()
}
\description{
Define \code{\link[=as_flextable]{as_flextable()}} as
print method in an R Markdown document for models
of class:
\itemize{
\item lm
\item glm
\item models from package 'lme' and 'lme4'
\item htest (t.test, chisq.test, ...)
\item gam
\item kmeans and pam
}

In a setup run chunk:\preformatted{flextable::use_model_printer()
}
}
\seealso{
\code{\link[=use_df_printer]{use_df_printer()}}, \code{\link[=flextable]{flextable()}}
}
