% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{void}
\alias{void}
\title{Delete flextable content}
\usage{
void(x, i = NULL, j = NULL, part = "body")
}
\arguments{
\item{x}{\code{flextable} object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{part}{partname of the table}
}
\description{
Deletion of flextable content. Values will
be replaced with \code{NA} values.
}
\examples{
ft <- flextable(mtcars)
ft <- void(ft, ~ drat > 3.5, ~ vs + am + gear + carb )
}
