% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{wml_flextable}
\alias{wml_flextable}
\title{wml table code}
\usage{
wml_flextable(x, relationships, standalone = TRUE)
}
\arguments{
\item{x}{\code{flextable} object}

\item{relationships}{relationships dataset}

\item{standalone}{specify to produce a standalone XML file.
If FALSE, omits xml header and default namespace.}
}
\description{
produces the wml of a flextable
}

