% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{write_docx}
\alias{write_docx}
\title{Microsoft Word table}
\usage{
write_docx(file, x)
}
\arguments{
\item{file}{filename of the Microsoft Word document to produce. File
extension must be \code{.docx}.}

\item{x}{flextable}
}
\description{
Table for Microsoft Word documents.
}
\examples{
ft <- flextable(head(iris))
write_docx(file = "test.docx", x = ft )
}

