% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{align}
\alias{align}
\title{Set text alignment}
\usage{
align(x, i = NULL, j = NULL, align = "left", part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{align}{text alignment - a single character value, expected value
is one of 'left', 'right', 'center', 'justify'.}

\item{part}{partname of the table (one of 'all', 'body', 'header')}
}
\description{
change text alignment of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- align(ft, align = "center")
}
