% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.pam}
\alias{as_flextable.pam}
\title{tabular summary for pam}
\usage{
\method{as_flextable}{pam}(x, digits = 4, ...)
}
\arguments{
\item{x}{a \code{\link[cluster:pam]{cluster::pam()}} object}

\item{digits}{number of digits for the numeric columns}

\item{...}{unused argument}
}
\description{
produce a flextable describing a
pam object. The function is only using package 'broom'
that provides the data presented in the resulting flextable.
}
\examples{
if(require("cluster")){
  dat <- as.data.frame(scale(mtcars[1:7]))
  cl <- pam(dat, 3)
  ft <- as_flextable(cl)
  ft
}
}
