% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable}
\alias{as_flextable}
\title{method to convert object to flextable}
\usage{
as_flextable(x, ...)
}
\arguments{
\item{x}{object to be transformed as flextable}

\item{...}{arguments for custom methods}
}
\description{
This is a convenient function
to let users create flextable bindings
from any objects.
}
\seealso{
Other as_flextable methods: 
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.xtable}()}
}
\concept{as_flextable methods}
