% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable}
\alias{as_flextable}
\title{Method to transform objects into flextables}
\usage{
as_flextable(x, ...)
}
\arguments{
\item{x}{object to be transformed as flextable}

\item{...}{arguments for custom methods}
}
\description{
This is a convenient function
to let users create flextable bindings
from any objects. Users should consult documentation
of corresponding method to understand the details and
see what arguments can be used.
}
\seealso{
Other as_flextable methods: 
\code{\link{as_flextable.data.frame}()},
\code{\link{as_flextable.gam}()},
\code{\link{as_flextable.glm}()},
\code{\link{as_flextable.grouped_data}()},
\code{\link{as_flextable.htest}()},
\code{\link{as_flextable.kmeans}()},
\code{\link{as_flextable.lm}()},
\code{\link{as_flextable.merMod}()},
\code{\link{as_flextable.pam}()},
\code{\link{as_flextable.summarizor}()},
\code{\link{as_flextable.tabular}()},
\code{\link{as_flextable.tabulator}()},
\code{\link{as_flextable.xtable}()}
}
\concept{as_flextable methods}
