% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{bg}
\alias{bg}
\title{Set background color}
\usage{
bg(x, i = NULL, j = NULL, bg, part = "body", source = j)
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{bg}{color to use as background color. If a function, function need to return
a character vector of colors.}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}

\item{source}{if bg is a function, source is specifying the dataset column to be used
as argument to \code{bg}. This is only useful if j is colored with values contained in another
(or other) column.}
}
\description{
change background color of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- bg(ft, bg = "#DDDDDD", part = "header")

if(require("scales")){
  ft <- flextable(iris)
  zz <- col_numeric(
    palette = c("red", "orange", "green", "blue"),
    domain = c(0, 7))
  ft <- bg(ft, j = c("Sepal.Length", "Sepal.Width",
                     "Petal.Length", "Petal.Width"),
           bg = zz, part = "body")
  ft
}

}
\seealso{
Other sugar functions for table style: 
\code{\link{align}()},
\code{\link{bold}()},
\code{\link{color}()},
\code{\link{empty_blanks}()},
\code{\link{fontsize}()},
\code{\link{font}()},
\code{\link{italic}()},
\code{\link{padding}()},
\code{\link{rotate}()},
\code{\link{valign}()}
}
\concept{sugar functions for table style}
