% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_flextable.R
\name{body_add_flextable}
\alias{body_add_flextable}
\title{wml table code}
\usage{
body_add_flextable(x, value, align = "center", pos = "after")
}
\arguments{
\item{x}{a docx object}

\item{value}{\code{flextable} object}

\item{align}{left (default), center or right.}

\item{pos}{where to add the flextable relative to the cursor,
one of "after", "before", "on" (end of line).}
}
\description{
produces the wml of a flextable
}
\examples{
library(officer)
ft <- flextable(head(mtcars))
ft <- theme_zebra(ft)
\donttest{ft <- autofit(ft)}
doc <- read_docx()
doc <- body_add_flextable(doc, value = ft)
print(doc, target = "test.docx")
}
