% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_docx.R
\name{body_add_flextable}
\alias{body_add_flextable}
\title{wml table code}
\usage{
body_add_flextable(x, value, pos = "after")
}
\arguments{
\item{x}{a docx object}

\item{value}{\code{flextable} object}

\item{pos}{where to add the flextable relative to the cursor,
one of "after", "before", "on" (end of line).}
}
\description{
produces the wml of a flextable
}
