% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{bold}
\alias{bold}
\title{Set font weight}
\usage{
bold(x, i = NULL, j = NULL, bold = TRUE, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{bold}{boolean value}

\item{part}{partname of the table (one of 'all', 'body', 'header')}
}
\description{
change font weight of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- bold(ft, bold = TRUE, part = "header")
}

