% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{border}
\alias{border}
\title{Set cell borders}
\usage{
border(x, i = NULL, j = NULL, border = NULL, border.top = NULL,
  border.bottom = NULL, border.left = NULL, border.right = NULL,
  part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{border}{border (shortcut for top, bottom, left and right)}

\item{border.top}{border top}

\item{border.bottom}{border bottom}

\item{border.left}{border left}

\item{border.right}{border right}

\item{part}{partname of the table (one of 'all', 'body', 'header')}
}
\description{
change borders of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- border(ft, border.top = fp_border(color = "orange") )
}
