% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders.R
\name{border_remove}
\alias{border_remove}
\title{remove borders}
\usage{
border_remove(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
The function is deleting all borders of the flextable object.
}
\examples{
dat <- iris[c(1:5, 51:55, 101:105),]
ft <- flextable(dat)
ft <- theme_box(ft)
ft

# remove all borders
ft <- border_remove(x = ft)
ft
}
\seealso{
Other borders management: 
\code{\link{border_inner_h}()},
\code{\link{border_inner_v}()},
\code{\link{border_inner}()},
\code{\link{border_outer}()},
\code{\link{border}()},
\code{\link{hline_bottom}()},
\code{\link{hline_top}()},
\code{\link{hline}()},
\code{\link{vline_left}()},
\code{\link{vline_right}()},
\code{\link{vline}()}
}
\concept{borders management}
