% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_rows.R
\name{delete_part}
\alias{delete_part}
\title{Delete flextable part}
\usage{
delete_part(x, part = "header")
}
\arguments{
\item{x}{a \code{flextable} object}

\item{part}{partname of the table to delete (one of 'body', 'header' or 'footer').}
}
\description{
indicate to not print a part of
the flextable, i.e. an header, footer or the body.
}
\examples{
ft <- flextable(head(iris))
ft <- delete_part(x = ft, part = "header")
ft
}
