% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim.flextable}
\alias{dim.flextable}
\title{Get widths and heights of flextable}
\usage{
\method{dim}{flextable}(x)
}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}
}
\description{
returns widths and heights for each table columns and rows.
Values are expressed in inches.
}
\examples{
ftab <- flextable(head(iris))
dim(ftab)
}
\seealso{
Other functions for flextable size management: 
\code{\link{autofit}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{ncol_keys}()},
\code{\link{nrow_part}()},
\code{\link{width}()}
}
\concept{functions for flextable size management}
