% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim.flextable}
\alias{dim.flextable}
\title{Get widths and heights of flextable}
\usage{
\method{dim}{flextable}(x)
}
\arguments{
\item{x}{flextable object}
}
\description{
returns widths and heights for each table columns and rows.
Values are expressed in inches.
}
\examples{
ft <- flextable(head(iris))
dim(ft)
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
