% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim.flextable}
\alias{dim.flextable}
\title{Get flextable dimensions}
\usage{
\method{dim}{flextable}(x)
}
\arguments{
\item{x}{flextable object}
}
\description{
returns widths and heights for each table columns and rows.
Values are inches.
}
