% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim_pretty}
\alias{dim_pretty}
\title{Calculate pretty dimensions}
\usage{
dim_pretty(x, part = "all", unit = "in")
}
\arguments{
\item{x}{flextable object}

\item{part}{partname of the table (one of 'all', 'body', 'header' or 'footer')}

\item{unit}{unit for returned values, one of "in", "cm", "mm".}
}
\description{
return minimum estimated widths and heights for
each table columns and rows in inches.
}
\section{line breaks}{

Soft returns (a line break in a paragraph) are not supported. Function
\code{dim_pretty} will return wrong results if \verb{\\n} are used (they will be
considered as "").
}

\examples{
ftab <- flextable(head(mtcars))
dim_pretty(ftab)
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{ncol_keys}()},
\code{\link{nrow_part}()},
\code{\link{set_table_properties}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
