% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{dim_pretty}
\alias{dim_pretty}
\title{Calculate pretty dimensions}
\usage{
dim_pretty(x, part = "all")
}
\arguments{
\item{x}{flextable object}

\item{part}{partname of the table (one of 'all', 'body', 'header' or 'footer')}
}
\description{
return minimum estimated widths and heights for
each table columns and rows in inches.
}
\examples{

ft <- flextable(mtcars)
\donttest{dim_pretty(ft)}
}
