% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{display}
\alias{display}
\title{flextable display values}
\usage{
display(x, i = NULL, part = "body", ...)
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{part}{partname of the table (one of 'all', 'body', 'header')}

\item{...}{see details.}
}
\description{
Modify flextable displayed values.
}
\details{
Use format_that or format_simple to define cell content.
}
\examples{

# Formatting data values example ------
ft <- flextable(head( mtcars, n = 10))
ft <- display(ft, i = ~ drat > 3.5,
  carb = fpar("# ", ftext(carb, fp_text(color="orange") ) ) )
write_docx("format_ft.docx", ft)
}
\seealso{
\code{\link{flextable}}
}

