% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{docx_value}
\alias{docx_value}
\title{flextable docx string}
\usage{
docx_value(x, print = TRUE)
}
\arguments{
\item{x}{a flextable object}

\item{print}{print output if TRUE}
}
\description{
get openxml raw code for Word
from a flextable object.
}
\examples{
docx_value(flextable(iris[1:5,]))
}
\seealso{
Other flextable print function: 
\code{\link{as_raster}()},
\code{\link{format.flextable}()},
\code{\link{htmltools_value}()},
\code{\link{knit_print.flextable}()},
\code{\link{plot.flextable}()},
\code{\link{print.flextable}()},
\code{\link{save_as_docx}()},
\code{\link{save_as_html}()},
\code{\link{save_as_image}()},
\code{\link{save_as_pptx}()}
}
\concept{flextable print function}
