% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{fit_to_width}
\alias{fit_to_width}
\title{fit a flextable to a maximum width}
\usage{
fit_to_width(x, max_width, inc = 1L, max_iter = 20)
}
\arguments{
\item{x}{flextable object}

\item{max_width}{maximum width to fit in inches}

\item{inc}{the font size decrease for each step}

\item{max_iter}{maximum iterations}
}
\description{
decrease font size for each cell incrementally until
it fits a given max_width.
}
\examples{
ft <- qflextable(head(mtcars))
ft <- padding(ft, padding = 0, part = "all")
fit_to_width(ft, max_width = 6)
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
