% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{flextable-defunct}
\alias{flextable-defunct}
\alias{ph_with_flextable}
\title{Defunct Functions in Package flextable}
\usage{
ph_with_flextable(...)
}
\arguments{
\item{...}{unused arguments}
}
\description{
Defunct Functions in Package flextable
}
\details{
\code{ph_with_flextable()} is replaced by \code{officer::ph_with}.
}
