% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_flextable_package.R
\docType{package}
\name{flextable-package}
\alias{flextable-package}
\title{flextable: Functions for Tabular Reporting}
\description{
The flextable package facilitates access to and manipulation of
tabular reporting elements from R.

The documentation of functions can be opened with command \code{help(package = "flextable")}.

To learn more about flextable, start with the vignettes:
\code{browseVignettes(package = "flextable")}.

\code{flextable()} function is producing flexible tables where each cell
can contain several chunks of text with their own set of formatting
properties (bold, font color, etc.). Function \code{\link[=compose]{compose()}} lets customise
text of cells.
}
\seealso{
\url{https://davidgohel.github.io/flextable/}, \code{\link[=flextable]{flextable()}}
}
