% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{flextable}
\alias{dim.flextable}
\alias{flextable}
\alias{optim_dim}
\alias{print.flextable}
\title{flextable object}
\usage{
flextable(data, col_keys = names(data))

\method{print}{flextable}(x, ...)

optim_dim(x)

\method{dim}{flextable}(x)
}
\arguments{
\item{data}{dataset}

\item{col_keys}{columns names/keys to display. If some column names are not in
the dataset, they will be added as blank columns.}

\item{x}{flextable object}

\item{...}{unused}
}
\description{
Create a flextable object with function \code{flextable}.

\code{flextable} are designed to make tabular reporting easier for
R users.

An API lets you format text, paragraphs and cells, table elements can
be merge vertically or horizontally.

A \code{flextable} is made of 2 parts: header and body.

Function \code{optim_dim} will return minimum estimated widths and heights for
each table columns and rows.

Function \code{dim} will return widths and heights for each table columns and rows.
}
\examples{
ft <- flextable(mtcars)
write_docx("ft.docx", ft)

# get estimated widths
ft <- flextable(mtcars)
optim_dim(ft)
}

