% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{flextable}
\alias{flextable}
\title{Create a flextable object}
\usage{
flextable(data, col_keys = names(data))
}
\arguments{
\item{data}{dataset}

\item{col_keys}{columns names/keys to display. If some column names are not in
the dataset, they will be added as blank columns.}
}
\description{
Create a flextable object with function \code{flextable}.

\code{flextable} are designed to make tabular reporting easier for
R users.

An API lets you format text, paragraphs and cells, table elements can
be merge vertically or horizontally.

A \code{flextable} is made of 2 parts: header and body.
}
\examples{
ft <- flextable(mtcars)
ft
}

