% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{flextable_dim}
\alias{flextable_dim}
\title{width and height of a flextable object}
\usage{
flextable_dim(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Returns the width, height and
aspect ratio of a flextable in a named list.
The width and height are in inches. The aspect ratio
is the ratio corresponding to \code{height/width}.

Names of the list are \code{width}, \code{height} and \code{aspect_ratio}.
}
\examples{
ft <- flextable(head(iris))
flextable_dim(ft)
ft <- autofit(ft)
flextable_dim(ft)
}
\seealso{
Other flextable dimensions: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{height}()},
\code{\link{width}()}
}
\concept{flextable dimensions}
