% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizor.R
\name{fmt_2stats}
\alias{fmt_2stats}
\title{format content for data generated with \code{\link[=summarizor]{summarizor()}}}
\usage{
fmt_2stats(
  num1,
  num2,
  cts,
  pcts,
  num1_mask = "\%.01f",
  num2_mask = "(\%.01f)",
  cts_mask = "\%.0f",
  pcts_mask = "(\%.02f \%\%)"
)
}
\arguments{
\item{num1}{a numeric statistic to display such as a mean or a median}

\item{num2}{a numeric statistic to display such as a standard
deviation or a median absolute deviation.}

\item{cts}{a count to display}

\item{pcts}{a percentage to display}

\item{num1_mask}{format associated with \code{num1}, a format string
used by \code{\link[=sprintf]{sprintf()}}.}

\item{num2_mask}{format associated with \code{num2}, a format string
used by \code{\link[=sprintf]{sprintf()}}.}

\item{cts_mask}{format associated with \code{cts}, a format string
used by \code{\link[=sprintf]{sprintf()}}.}

\item{pcts_mask}{format associated with \code{pcts}, a format string
used by \code{\link[=sprintf]{sprintf()}}.}
}
\description{
This function was written to allow easy demonstrations
of flextable's ability to produce table summaries (with \code{\link[=summarizor]{summarizor()}}).
It assumes that we have either a quantitative variable, in which
case we will display the mean and the standard deviation, or a
qualitative variable, in which case we will display the count and the
percentage corresponding to each modality.
}
\seealso{
\code{\link[=summarizor]{summarizor()}}, \code{\link[=tabulator]{tabulator()}}, \code{\link[=mk_par]{mk_par()}}
}
