% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizor.R
\name{fmt_n_percent}
\alias{fmt_n_percent}
\title{format content for count data}
\usage{
fmt_n_percent(n, pct, digit = 1)
}
\arguments{
\item{n}{count values}

\item{pct}{percent values}

\item{digit}{number of digits for the percentages}
}
\description{
The function formats counts and
percentages as \verb{n (xx.x\%)}. If percentages are
missing, they are not printed.
}
\seealso{
\code{\link[=tabulator]{tabulator()}}, \code{\link[=mk_par]{mk_par()}}

Other other text formatter functions: 
\code{\link{fmt_2stats}()},
\code{\link{fmt_header_n}()}
}
\concept{other text formatter functions}
