% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{font}
\alias{font}
\title{Set font}
\usage{
font(x, i = NULL, j = NULL, fontname, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{fontname}{string value, the font name.}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
change font of selected rows and columns of a flextable.
}
\examples{
require("gdtools")
fontname <- "Times"

if( !font_family_exists(fontname) ){
  ft <- flextable(head(iris))
  ft <- font(ft, fontname = fontname, part = "header")
}
}
\seealso{
Other sugar functions for table style: \code{\link{align}},
  \code{\link{bg}}, \code{\link{bold}},
  \code{\link{color}}, \code{\link{empty_blanks}},
  \code{\link{fontsize}}, \code{\link{italic}},
  \code{\link{padding}}, \code{\link{rotate}},
  \code{\link{valign}}
}
\concept{sugar functions for table style}
