% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{fontsize}
\alias{fontsize}
\title{Set font size}
\usage{
fontsize(x, i = NULL, j = NULL, size = 11, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{size}{integer value (points)}

\item{part}{partname of the table (one of 'all', 'body', 'header')}
}
\description{
change font size of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- fontsize(ft, size = 14, part = "header")
}
