% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_add_flextable.R
\name{footers_flextable_at_bkm}
\alias{footers_flextable_at_bkm}
\title{add flextable at a bookmark location in document's footer}
\usage{
footers_flextable_at_bkm(x, bookmark, value)
}
\arguments{
\item{x}{an rdocx object}

\item{bookmark}{bookmark id}

\item{value}{a flextable object}
}
\description{
replace in the footer of a document  a paragraph containing a bookmark by a flextable.
A bookmark will be considered as valid if enclosing words
within a paragraph; i.e., a bookmark along two or more paragraphs is invalid,
a bookmark set on a whole paragraph is also invalid, but bookmarking few words
inside a paragraph is valid.
}
