% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{fp_border_default}
\alias{fp_border_default}
\title{Border formatting properties}
\usage{
fp_border_default(
  color = flextable_global$defaults$border.color,
  style = "solid",
  width = 1
)
}
\arguments{
\item{color}{border color - single character value (e.g. "#000000" or "black")}

\item{style}{border style - single character value : "none" or "solid" or "dotted" or "dashed"}

\item{width}{border width - an integer value : 0>= value}
}
\description{
Create a \code{\link[=fp_border]{fp_border()}} object that uses
defaut values defined in flextable defaults formatting properties, i.e.
default border color (see \code{\link[=set_flextable_defaults]{set_flextable_defaults()}}).
}
\examples{
fp_border_default(width = 2)
}
\seealso{
\code{\link[=hline]{hline()}}, \code{\link[=vline]{vline()}}

Other functions for defining formatting properties: 
\code{\link{fp_text_default}()}
}
\concept{functions for defining formatting properties}
