% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{height}
\alias{height}
\title{Set flextable rows height}
\usage{
height(x, i = NULL, height, part = "body")
}
\arguments{
\item{x}{flextable object}

\item{i}{rows selection}

\item{height}{height in inches}

\item{part}{partname of the table}
}
\description{
control rows height
}
\examples{

ft <- flextable(iris)
ft <- height(ft, height = .3)

write_docx("height.docx", ft)
}
