% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{hline}
\alias{hline}
\alias{hline_top}
\alias{hline_bottom}
\alias{vline_left}
\alias{vline_right}
\alias{vline}
\title{Set horizontal or vertical borders}
\usage{
hline(x, i = NULL, j = NULL, border = NULL, part = "body")

hline_top(x, j = NULL, border = NULL, part = "body")

hline_bottom(x, j = NULL, border = NULL, part = "body")

vline_left(x, i = NULL, border = NULL, part = "body")

vline_right(x, i = NULL, border = NULL, part = "body")

vline(x, i = NULL, j = NULL, border = NULL, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{border}{border}

\item{part}{partname of the table (one of 'all', 'body', 'header', 'footer')}
}
\description{
change horizontal or vertical borders of a
flextable (bottom side). These functions are taking care
of propagating the border properties to the adjacent cells.
}
\examples{
library(officer)
big_border = fp_border(color="gray", width = 2)
std_border = fp_border(color="orange", width = 1)

dat <- iris[c(1:5, 51:55, 101:105),]
ft <- regulartable(dat, col_keys = c("Species", "Sepal.Length",
  "Sepal.Width", "Petal.Length", "Petal.Width" ))
ft <- border(x = ft, part="all", border =fp_border(color="transparent"))
ft <- merge_v(ft, j = 1)
ft <- hline(ft, part="all", border = std_border )
ft <- hline(ft, part="header", i = 1, border = big_border )
ft <- hline(ft, part="body", i = c(5, 10, 15), border = big_border )
ft <- hline_top(ft, part="header", border = big_border )
ft <- vline(ft, part = "all", border = std_border)
ft <- vline_left(ft, part = "all", border = std_border)
ft
}
