% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{htmltools_value}
\alias{htmltools_value}
\title{flextable a tag object from htmltools package}
\usage{
htmltools_value(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
get a \code{\link[htmltools]{div}} from a flextable object.
This can be used in a shiny application.
}
\examples{
htmltools_value(flextable(iris[1:5,]))
}
\seealso{
Other flextable print function: \code{\link{as_raster}},
  \code{\link{format.flextable}},
  \code{\link{knit_print.flextable}},
  \code{\link{plot.flextable}},
  \code{\link{print.flextable}},
  \code{\link{save_as_html}}, \code{\link{save_as_image}}
}
\concept{flextable print function}
