% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{italic}
\alias{italic}
\title{italic font}
\usage{
italic(x, i = NULL, j = NULL, italic = TRUE, part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{i}{rows selection}

\item{j}{columns selection}

\item{italic}{boolean value}

\item{part}{partname of the table (one of 'all', 'body', 'header')}
}
\description{
change font decoration of selected rows and columns of a flextable.
}
\examples{
ft <- flextable(mtcars)
ft <- italic(ft, italic = TRUE, part = "header")
}

