% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{knit_print.flextable}
\alias{knit_print.flextable}
\title{Render flextable in rmarkdown (including Word output)}
\usage{
\method{knit_print}{flextable}(x, ...)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{...}{further arguments, not used.}
}
\description{
Function used to render flextable in knitr/rmarkdown documents.
HTML and Word outputs are supported.
}
\note{
For Word (docx) output, if pandoc vesion >= 2.0 is used, a raw XML block
with the table code will be inserted. If pandoc vesion < 2.0 is used, an
error will be raised. Note also that insertion of images is not supported
with rmarkdow for Word documents.
}
\author{
Maxim Nazarov
}
