% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrowcol.R
\name{nrow_part}
\alias{nrow_part}
\title{Number of rows of a part}
\usage{
nrow_part(x, part = "body")
}
\arguments{
\item{x}{a 'flextable' object, see \link{flextable-package} to learn how to create
'flextable' object.}

\item{part}{part selector, see section \emph{Part selection with the \code{part}
parameter} in <\code{\link[=flextable_selectors]{Selectors in flextable}}>.
Value 'all' can be used.}
}
\description{
returns the number of lines in a part of flextable.
}
\examples{
library(flextable)
ft <- qflextable(head(cars))
nrow_part(ft, part = "body")
}
\seealso{
Other functions for flextable size management: 
\code{\link{autofit}()},
\code{\link{dim.flextable}()},
\code{\link{dim_pretty}()},
\code{\link{fit_to_width}()},
\code{\link{flextable_dim}()},
\code{\link{height}()},
\code{\link{hrule}()},
\code{\link{ncol_keys}()},
\code{\link{set_table_properties}()},
\code{\link{width}()}
}
\concept{functions for flextable size management}
