% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ph_with_flextable}
\alias{ph_with_flextable}
\title{add flextable into a PowerPoint slide}
\usage{
ph_with_flextable(x, value, type = "body", index = 1)
}
\arguments{
\item{x}{an rpptx device}

\item{value}{\code{flextable} object}

\item{type}{placeholder type}

\item{index}{placeholder index (integer). This is to be used when a placeholder type
is not unique in the current slide, e.g. two placeholders with type 'body'.}
}
\description{
add a flextable as a new shape in the current slide.

These functions is deprecated and
method \code{\link[officer]{ph_with}} should be used instead.
}
\note{
The width and height of the table can not be set with this function. Use
functions \code{\link{width}}, \code{\link{height}}, \code{\link{autofit}}
and \code{\link{dim_pretty}} instead. The overall size is resulting from
cells, paragraphs and text properties (i.e. padding, font size, border widths).
}
