% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printers.R
\name{plot.flextable}
\alias{plot.flextable}
\title{plot a flextable}
\usage{
\method{plot}{flextable}(x, zoom = 2, expand = 2, ...)
}
\arguments{
\item{x}{a flextable object}

\item{zoom, expand}{parameters used by \code{webshot} function.}

\item{...}{additional parameters sent to plot function}
}
\description{
save a flextable as an image and display the
result in a new R graphics window.
}
\note{
This function requires packages: webshot and magick.
}
\examples{
ft <- flextable( head( mtcars ) )
ft <- autofit(ft)
\dontrun{
if( require("webshot") ){
  plot(ft)
}
}
}
\seealso{
Other flextable print function: \code{\link{as_raster}},
  \code{\link{format.flextable}},
  \code{\link{htmltools_value}},
  \code{\link{knit_print.flextable}},
  \code{\link{print.flextable}},
  \code{\link{save_as_html}}, \code{\link{save_as_image}}
}
\concept{flextable print function}
