% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{set_caption}
\alias{set_caption}
\title{set caption}
\usage{
set_caption(
  x,
  caption,
  autonum = NULL,
  ref = NULL,
  style = "Table Caption",
  html_escape = TRUE
)
}
\arguments{
\item{x}{flextable object}

\item{caption}{caption value}

\item{autonum}{an autonum representation. See \code{\link[officer]{run_autonum}}.
This has only an effect when output is Word. If used, the caption is preceded
by an auto-number sequence. In this case, the caption is preceded by an auto-number
sequence.}

\item{ref}{a single character identifier that will be used as bookmark label.
This has only an effect when output is Word. If used, the caption is bookmarked
and can be cross-referenced. See also \code{\link[officer]{run_reference}}.}

\item{style}{caption paragraph style name. These names are available with
function \code{\link[officer]{styles_info}}.}

\item{html_escape}{should HTML entities be escaped so that it can be safely
included as text or an attribute value within an HTML document.}
}
\description{
set caption value in flextable
}
\note{
this will have an effect only when output is HTML or Word document.
}
\examples{
ftab <- flextable( head( iris ) )
ftab <- set_caption(ftab, "my caption")
ftab
}
