% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{style}
\alias{bg}
\alias{bold}
\alias{border}
\alias{color}
\alias{padding}
\alias{style}
\title{flextable style}
\usage{
style(x, ..., i = NULL, j = NULL, part = "body")

bg(x, i = NULL, j = NULL, bg, part = "body")

bold(x, i = NULL, j = NULL, bold = TRUE, part = "body")

color(x, i = NULL, j = NULL, color, part = "body")

padding(x, i = NULL, j = NULL, padding = NULL, padding.top = NULL,
  padding.bottom = NULL, padding.left = NULL, padding.right = NULL,
  part = "body")

border(x, i = NULL, j = NULL, border = NULL, border.top = NULL,
  border.bottom = NULL, border.left = NULL, border.right = NULL,
  part = "body")
}
\arguments{
\item{x}{a flextable object}

\item{...}{object(s) of class \code{pr_text}, \code{pr_par} and \code{pr_cell}.}

\item{i}{rows selection}

\item{j}{columns selection}

\item{part}{partname of the table (one of 'all', 'body', 'header')}

\item{bg}{color to use as background color}

\item{bold}{boolean value}

\item{color}{color to use as font color}

\item{padding}{padding (shortcut for top, bottom, left and right)}

\item{padding.top}{padding top}

\item{padding.bottom}{padding bottom}

\item{padding.left}{padding left}

\item{padding.right}{padding right}

\item{border}{border (shortcut for top, bottom, left and right)}

\item{border.top}{border top}

\item{border.bottom}{border bottom}

\item{border.left}{border left}

\item{border.right}{border right}
}
\description{
Modify flextable text, paragraphs and cells formatting properties with function \code{style}.

Set background color with function \code{bg}.

Set font weight to bold with function \code{bold}.

Set text color with function \code{color}.

Set paragraphs paddings with function \code{padding}.

Set cell borders with function \code{border}.
}
\examples{

# Styles example ------
def_cell <- pr_cell(border = pr_border(color="#00FFFF"))

def_par <- pr_par(text.align = "center")

ft <- flextable(mtcars)

ft <- style( ft, def_cell, def_par, parts = "all")
ft <- style(ft, ~ drat > 3.5, ~ vs + am + gear + carb,
  pr_text(color="red", italic = TRUE) )

write_docx("style_ft.docx", ft)

# bg example ------
ft <- flextable(mtcars)
ft <- bg(ft, bg = "#DDDDDD", part = "header")

# bold example ------
ft <- flextable(mtcars)
ft <- bold(ft, bold = TRUE, part = "header")

# color example ------
ft <- flextable(mtcars)
ft <- color(ft, color = "orange", part = "header")

# padding example ------
ft <- flextable(mtcars)
ft <- padding(ft, padding.top = 4)

# border example ------
ft <- flextable(mtcars)
ft <- border(ft, border.top = pr_border(color = "orange") )
}

