% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabwid.R
\name{tabwid}
\alias{tabwid}
\title{<Add Title>}
\usage{
tabwid(x, width = NULL, height = NULL)
}
\arguments{
\item{x}{\code{flextable} object}

\item{width}{widget width}

\item{height}{widget height}
}
\description{
<Add Description>
}

