% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabwid.R
\name{tabwid}
\alias{tabwid}
\title{htmlwidget for flextable}
\usage{
tabwid(x, width = NULL, height = NULL)
}
\arguments{
\item{x}{\code{flextable} object}

\item{width}{widget width. Unused argument, columns widths are managed by function width.}

\item{height}{widget height Unused argument, rows heights are managed by function height.}
}
\description{
htmlwidget for flextable
}

