% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_booktabs}
\alias{theme_booktabs}
\title{Apply booktabs theme}
\usage{
theme_booktabs(x, fontsize = 11)
}
\arguments{
\item{x}{a flextable object}

\item{fontsize}{font size in pixel}
}
\description{
Apply theme booktabs to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_booktabs(ft)
}
\seealso{
Other flextable theme: 
\code{\link{theme_alafoli}()},
\code{\link{theme_box}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{flextable theme}
