% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_box}
\alias{theme_box}
\title{Apply box theme}
\usage{
theme_box(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme box to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_box(ft)
}
