% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_box}
\alias{theme_box}
\title{Apply box theme}
\usage{
theme_box(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme box to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_box(ft)
}
\seealso{
Other flextable theme: 
\code{\link{theme_alafoli}()},
\code{\link{theme_booktabs}()},
\code{\link{theme_tron_legacy}()},
\code{\link{theme_tron}()},
\code{\link{theme_vader}()},
\code{\link{theme_vanilla}()},
\code{\link{theme_zebra}()}
}
\concept{flextable theme}
