% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tron}
\alias{theme_tron}
\title{Apply tron theme}
\usage{
theme_tron(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme tron to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_tron(ft)
}
