% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_tron_legacy}
\alias{theme_tron_legacy}
\title{Apply tron legacy theme}
\usage{
theme_tron_legacy(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme tron legacy to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_tron_legacy(ft)
}
