% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_vanilla}
\alias{theme_vanilla}
\title{Apply vanilla theme}
\usage{
theme_vanilla(x)
}
\arguments{
\item{x}{a flextable object}
}
\description{
Apply theme vanilla to a flextable
}
\examples{
ft <- flextable(iris)
ft <- theme_vanilla(ft)
}
