% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{void}
\alias{void}
\title{Delete flextable content}
\usage{
void(x, j = NULL, part = "body")
}
\arguments{
\item{x}{\code{flextable} object}

\item{j}{columns selection}

\item{part}{partname of the table}
}
\description{
Set content display as a blank \code{" "}.
}
\examples{
ftab <- flextable(head(mtcars))
ftab <- void(ftab, ~ vs + am + gear + carb )
ftab
}
