% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable_sizes.R
\name{width}
\alias{width}
\title{Set flextable columns width}
\usage{
width(x, j = NULL, width)
}
\arguments{
\item{x}{flextable object}

\item{j}{columns selection}

\item{width}{width in inches}
}
\description{
control columns width
}
\details{
Heights are not used when flextable is been rendered into HTML.
}
\examples{

ft <- flextable(iris)
ft <- width(ft, width = 1)

write_docx("autofit.docx", ft)
}
\seealso{
\code{\link{flextable}}
}
