% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{width}
\alias{width}
\title{set columns width}
\usage{
width(x, j = NULL, width)
}
\arguments{
\item{x}{a flextable object}

\item{j}{columns selection}

\item{width}{width in inches}
}
\description{
modify columns widths of a flextable
}
\examples{

ft <- vanilla_table(iris)
ft <- width(ft, width = 1)

write_docx("autofit.docx", ft)
}
\seealso{
\code{\link{flextable}}
}

