% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pptx.R
\name{write_pptx}
\alias{write_pptx}
\title{Microsoft PowerPoint table}
\usage{
write_pptx(file, x)
}
\arguments{
\item{file}{filename of the Microsoft PowerPoint document to produce. File
extension must be \code{.pptx}.}

\item{x}{flextable}
}
\description{
Table for Microsoft PowerPoint documents.
}
\examples{
ft <- flextable(head(iris))
write_pptx(file = "test.pptx", x = ft )
}

